Moved utility cf-check functions to libutils
