+++ libutils/Makefile.am
@@ -38,6 +38,7 @@ libutils_la_SOURCES = \
 	deprecated.h \
 	dir.h dir_priv.h \
 	hash_method.h \
+	path.c path.h \
 	sequence.c sequence.h \
 	set.c set.h \
 	statistics.c statistics.h \
