+++ libutils/sequence.h
@@ -267,4 +267,13 @@ char *SeqStringSerialize(Seq *seq);
  */
 Seq *SeqStringDeserialize(const char *serialized);
 
+/**
+ * @brief Convert the argc + argv in main function to a sequence of strings
+ *
+ * Does not copy, edit or free the elements in argv.
+ * Shallow copy. SeqDestroy is safe, it will not free the contents of argv.
+ *
+ */
+Seq *SeqFromArgv(int argc, char **argv);
+
 #endif
